Split / combine code into modules in src/
