+++ .editorconfig
@@ -2,7 +2,7 @@ root = true
 
 [*]
 end_of_line = lf
-insert_final_newline = false
+insert_final_newline = true
 
 [*.py]
 charset = utf-8
